refactor: move getFindingsSeverityTrends to severity-trends module
- Move getFindingsSeverityTrends implementation from overview.ts to severity-trends.ts
- Define function where it's actually used, improving code organization
- Remove getFindingsSeverityTrends from overview.ts module
- Update imports in SSR components to use severity-trends module
- Export getFindingsSeverityTrends from severity-trends for re-export
- Fix semantic color classes in error states
